% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMEXClasses.R
\name{coef.SIMEXResult}
\alias{coef.SIMEXResult}
\title{Provide the Coefficient (Parameter Estimates) of the Model}
\usage{
\method{coef}{SIMEXResult}(object)
}
\arguments{
\item{obj}{an instance of the S3 class SIMEXResult}
}
\value{
a named vector
}
\description{
Provide the Coefficient (Parameter Estimates) of the Model
}
