% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMEXGLM.R
\name{createDataSet}
\alias{createDataSet}
\title{Create a Data Structure}
\usage{
createDataSet(formula, data, varianceFieldName = NULL)
}
\arguments{
\item{formula}{a formula (e.g. "y ~ x")}

\item{data}{a data.frame object}

\item{varianceFieldName}{the field that contains the variance of the measurement error in the data argument}
}
\description{
Create a data structure on the Java end that will be later used with
the GLM or SIMEX methods.
}
