% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMEXGLM.R
\name{.SIMEXDataSet}
\alias{.SIMEXDataSet}
\title{Create a Data Structure for the SIMEX Method}
\usage{
.SIMEXDataSet(formula, data, varianceFieldName)
}
\arguments{
\item{formula}{a formula (e.g. "y ~ x")}

\item{data}{a data.frame object}

\item{varianceFieldName}{the field that contains the variance of the measurement error in the data argument}
}
\description{
Create a data structure on the Java end that will be later used with
the SIMEX method.
}
