% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMEXClasses.R
\name{fitted.SIMEXResult}
\alias{fitted.SIMEXResult}
\title{Provide the Predictions Based on the SIMEX Parameter Estimates}
\usage{
\method{fitted}{SIMEXResult}(object)
}
\arguments{
\item{object}{an instance of the S3 class SIMEXResult}
}
\value{
a vector
}
\description{
Provide the Predictions Based on the SIMEX Parameter Estimates
}
