% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMEXGLM.R
\name{GLM}
\alias{GLM}
\title{Fit Generalized Linear Models}
\usage{
GLM(
  formula,
  dist = c("Bernoulli", "NegativeBinomial"),
  linkFunction = c("Logit", "CLogLog", "Log"),
  data,
  debug = F
)
}
\arguments{
\item{formula}{a formula (e.g. "y ~ x")}

\item{dist}{the distribution of the reponse variable (either "Bernoulli" or "NegativeBinomial")}

\item{linkFunction}{the link function (either "Logit" or "CLogLog" for the Bernoulli distribution
or "Log" for the negative binomial)}

\item{data}{a data.frame object}

\item{debug}{a logical to enable the debug mode}
}
\value{
a java.object instance
}
\description{
First, it create a data structure on the Java end. Then,
it fits a generalized linear model.
}
