% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMEXClasses.R
\name{plot.SIMEXResult}
\alias{plot.SIMEXResult}
\title{Plot the Observed and Predicted Parameter Estimates}
\usage{
\method{plot}{SIMEXResult}(object)
}
\arguments{
\item{obj}{an instance of the S3 class SIMEXResult}
}
\description{
Plot the Observed and Predicted Parameter Estimates
}
