% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMEXClasses.R
\name{predict.SIMEXResult}
\alias{predict.SIMEXResult}
\title{Make Predictions Based on the SIMEX Parameter Estimate}
\usage{
\method{predict}{SIMEXResult}(object, newdata, originalFormula = object$formula)
}
\arguments{
\item{object}{a SIMEXResult instance}

\item{newdata}{a data.frame with new data}

\item{originalFormula}{a string representing the original formula}
}
\value{
a data.frame containing the predictions and their 0.95 confidence intervals
}
\description{
Make Predictions Based on the SIMEX Parameter Estimate
}
