% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMEXGLM.R
\docType{data}
\name{simexExample}
\alias{simexExample}
\title{A fake data.frame object for an example of the SIMEX method}
\format{
An object of class \code{data.frame} with 2000 rows and 3 columns.
}
\usage{
data(simexExample)
}
\description{
A fake data.frame object for an example of the SIMEX method
}
\examples{
data(simexExample)
}
\keyword{datasets}
