% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMEXGLM.R
\docType{data}
\name{simexExampleNegBinomial}
\alias{simexExampleNegBinomial}
\title{A fake data.frame object for an example of the SIMEX method
with negative binomial regression.}
\format{
An object of class \code{data.frame} with 214 rows and 7 columns.
}
\usage{
data(simexExampleNegBinomial)
}
\description{
A fake data.frame object for an example of the SIMEX method
with negative binomial regression.
}
\examples{
data(simexExampleNegBinomial)
}
\keyword{datasets}
