% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMEXClasses.R
\name{summary.SIMEXResult}
\alias{summary.SIMEXResult}
\title{Provide a Summary Adapted to the SIMEX Method}
\usage{
\method{summary}{SIMEXResult}(object)
}
\arguments{
\item{obj}{an instance of the S3 class SIMEXResult}
}
\description{
Provide a Summary Adapted to the SIMEX Method
}
