% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMEXClasses.R
\name{vcov.SIMEXResult}
\alias{vcov.SIMEXResult}
\title{Provide the Estimated Variance-Covaraince of the Model Coefficients}
\usage{
\method{vcov}{SIMEXResult}(object)
}
\arguments{
\item{obj}{an instance of the S3 class SIMEXResult}
}
\description{
Provide the Estimated Variance-Covaraince of the Model Coefficients
}
