% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMEXGLM.R
\name{SIMEXGLM}
\alias{SIMEXGLM}
\title{Correct for the Measurement using the SIMEX Method}
\usage{
SIMEXGLM(
  formula,
  linkFunction = c("logit", "CLogLog"),
  data,
  fieldWithMeasError,
  varianceFieldName,
  factors = seq(0, 2, by = 0.2),
  nbBootstrapRealizations = 100,
  nbThreads = 2,
  debug = F
)
}
\arguments{
\item{formula}{a formula (e.g. "y ~ x")}

\item{linkFunction}{the link function (either "logit" or "CLogLog")}

\item{data}{a data.frame object}

\item{fieldWithMeasError}{the field with measurement error in the data argument}

\item{varianceFieldName}{the field that contains the variance of the measurement error in the data argument}

\item{factors}{a sequence of inflation factors (from 0 to 2 by .2 by default)}

\item{nbBootstrapRealizations}{the number of bootstrap realizations for each level of inflated variance (is
set to 100 by default )}

\item{nbThreads}{the number of threads to process the bootstrap realizations (is set to 2 by default)}

\item{debug}{a logical to enable the debug mode}
}
\value{
an instance of the S3 SIMEXResult class
}
\description{
First, create a data structure on the Java end that will be later used with
the SIMEX method. Secondly, fit a naive model. Thirdly, implements the SIMEX
method.
}
